add base input component
